Bug 2027644 - Part 3: Add TimeClip with int64 input. r=iain
This avoids calling `std::trunc` through `JS::ToInteger` in `JS::TimeClip`.

Differential Revision: https://phabricator.services.mozilla.com/D290824

+++ dom/html/HTMLInputElement.cpp
@@ -1956,7 +1956,7 @@ void HTMLInputElement::GetValueAsDate(JSContext* aCx,
         return;
       }
 
-      time.emplace(JS::TimeClip(millisecond));
+      time.emplace(JS::TimeClip(int64_t(millisecond)));
       MOZ_ASSERT(time->toDouble() == millisecond,
                  "HTML times are restricted to the day after the epoch and "
                  "never clip");
